mmc: msm_sdcc: Add software command timeout support
There might be cases where the controller may be bad and the
interrupts might be missed due to which the requests will be
stalled indefinitely. Also some bad cards don't come out of
programming state holding the prog_done line asserted and
mmcqd thread stall indefinitely waiting for prog_done interrupt.

Handle these cases with software timeouts in the driver file.
Specification defines the timeout value to be much less than 10secs
(around 1 sec) but we set the software command timeout to this
value to handle the controller/card which don't meet the standards.

Change-Id: Ic4635ba89e92eb4712b3e55e0f9ff11e9c6c2466
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>