mmc: msm_sdcc: Check for card presence during system resume
SDC controllers which support hardware detection the interrupt
generated due to insertion of SD card during suspend will be missed
as the corresponding irq is disabled during suspend. The driver remembers
the old eject status of the GPIO line (before suspend) and do not allow
any new requests (to detect the card) after resume even if the card is
present. Fix this by checking the GPIO status during resume and update
the flags so that card detection can happen if the card is present
after resume.

CRs-Fixed: 276417
Change-Id: I0e677ef6b3269dccd3a70283c592ca629104da3f
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>

Conflicts:

	drivers/mmc/host/msm_sdcc.c